A+B subspace
